backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pandas==2.2.2
SQLAlchemy==2.0.35
python-dotenv==1.0.1
httpx==0.27.2
    